# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: LexFrame Stage 8 API
  version: 0.8.0
  description: Canonical HTTP contract for the LexFrame stage 8 execution core, embedded builder, approval inbox, artifact handling, delivery and notifications.
servers:
  - url: http://localhost:3100
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [Documents]
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: kind
          required: false
          schema:
            $ref: "#/components/schemas/DocumentKind"
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/DocumentStatus"
        - in: query
          name: classification
          required: false
          schema:
            type: string
            enum:
              [
                public,
                internal,
                confidential,
                legal_secret,
                personal_data,
                client_material,
              ]
        - in: query
          name: tag
          required: false
          schema:
            type: string
        - in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
  /documents/upload-intents:
    post:
      operationId: createDocumentUploadIntent
      tags: [Documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUploadIntentRequest"
      responses:
        "200":
          description: Upload intent for a new document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadIntentResponse"
  /documents/{documentId}:
    get:
      operationId: getDocument
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Detailed document payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
    delete:
      operationId: deleteDocument
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document soft-delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMutationResult"
  /documents/{documentId}/versions:
    get:
      operationId: listDocumentVersions
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersionSummary"
  /documents/{documentId}/versions/upload-intent:
    post:
      operationId: createDocumentVersionUploadIntent
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentVersionUploadIntentRequest"
      responses:
        "200":
          description: Upload intent for a new document version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadIntentResponse"
  /documents/{documentId}/versions/{versionId}/complete:
    post:
      operationId: completeDocumentUpload
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/VersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUploadRequest"
      responses:
        "200":
          description: Document detail after upload completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
  /documents/{documentId}/versions/{versionId}/make-current:
    post:
      operationId: makeDocumentVersionCurrent
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/VersionId"
      responses:
        "200":
          description: Document detail after version switch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
  /documents/{documentId}/signed-url:
    post:
      operationId: createDocumentSignedUrl
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignedUrlRequest"
      responses:
        "200":
          description: Signed URL for the selected document asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedUrlResponse"
  /documents/{documentId}/archive:
    post:
      operationId: archiveDocument
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Archive result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMutationResult"
  /documents/{documentId}/restore:
    post:
      operationId: restoreDocument
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Restore result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMutationResult"
...
components:
  parameters:
...
    DocumentId:
      in: path
      name: documentId
      required: true
      schema:
        type: string
    VersionId:
      in: path
      name: versionId
      required: true
      schema:
        type: string
...
  schemas:
...
    DocumentKind:
      type: string
      enum:
        [
          case_material,
          evidence,
          legal_source,
          document_template,
          generated_document,
          draft_document,
          delivery_attachment,
          profile_clause,
          other,
        ]
    DocumentStatus:
      type: string
      enum:
        [
          upload_pending,
          uploaded,
          processing,
          ready,
          failed,
          archived,
          soft_deleted,
          hard_delete_pending,
        ]
    DocumentSource:
      type: string
      enum:
        [
          user_upload,
          automation_result,
          activepieces_artifact,
          ai_generated,
          template_library,
          profile_library,
          system_import,
        ]
    DocumentObjectRole:
      type: string
      enum: [original, preview_pdf, thumbnail, extracted_text, redacted_copy]
    DocumentJobType:
      type: string
      enum:
        [
          virus_scan,
          metadata_extract,
          text_extract,
          preview_generate,
          thumbnail_generate,
          index_prepare,
        ]
    DocumentJobStatus:
      type: string
      enum: [queued, running, completed, failed, skipped]
    DocumentScanStatus:
      type: string
      enum:
        [
          not_started,
          queued,
          clean,
          infected,
          manual_review_required,
          not_configured,
        ]
    DocumentPreviewStatus:
      type: string
      enum: [not_started, queued, ready, failed]
    DocumentExtractionStatus:
      type: string
      enum: [not_started, queued, ready, failed, requires_ocr]
    StorageState:
      type: string
      enum: [private_bucket, signed_url_only, quarantined, draft]
    UploadMethod:
      type: string
      enum: [direct, backend_proxy, service_ingest]
    DocumentVersionSummary:
      type: object
      required:
        [
          id,
          documentId,
          versionNo,
          status,
          originalFilename,
          mimeType,
          sizeBytes,
          sha256,
          storageState,
          scanStatus,
          previewStatus,
          extractionStatus,
          createdAt,
          completedAt,
        ]
      properties:
        id:
          type: string
        documentId:
          type: string
        versionNo:
          type: integer
        status:
          $ref: "#/components/schemas/DocumentStatus"
        originalFilename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        sha256:
          type: [string, "null"]
        storageState:
          $ref: "#/components/schemas/StorageState"
        scanStatus:
          $ref: "#/components/schemas/DocumentScanStatus"
        previewStatus:
          $ref: "#/components/schemas/DocumentPreviewStatus"
        extractionStatus:
          $ref: "#/components/schemas/DocumentExtractionStatus"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
    DocumentStorageObject:
      type: object
      required: [id, versionId, role, mimeType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
        versionId:
          type: string
        role:
          $ref: "#/components/schemas/DocumentObjectRole"
        mimeType:
          type: string
        sizeBytes:
          type: [integer, "null"]
        status:
          $ref: "#/components/schemas/StorageState"
        createdAt:
          type: string
          format: date-time
    DocumentRelation:
      type: object
      required: [id, relationType, targetEntityType, targetEntityId, createdAt]
      properties:
        id:
          type: string
        relationType:
          type: string
        targetEntityType:
          type: string
        targetEntityId:
          type: string
        createdAt:
          type: string
          format: date-time
    DocumentProcessingJob:
      type: object
      required:
        [
          id,
          versionId,
          jobType,
          status,
          attempts,
          maxAttempts,
          lastError,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        versionId:
          type: string
        jobType:
          $ref: "#/components/schemas/DocumentJobType"
        status:
          $ref: "#/components/schemas/DocumentJobStatus"
        attempts:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentSummary:
      type: object
      required:
        [
          id,
          workspaceId,
          ownerId,
          title,
          description,
          kind,
          status,
          classification,
          source,
          tags,
          currentVersion,
          createdAt,
          updatedAt,
          archivedAt,
          deletedAt,
        ]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        ownerId:
          type: string
        title:
          type: string
        description:
          type: [string, "null"]
        kind:
          $ref: "#/components/schemas/DocumentKind"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        classification:
          type: string
          enum:
            [
              public,
              internal,
              confidential,
              legal_secret,
              personal_data,
              client_material,
            ]
        source:
          $ref: "#/components/schemas/DocumentSource"
        tags:
          type: array
          items:
            type: string
        currentVersion:
          oneOf:
            - $ref: "#/components/schemas/DocumentVersionSummary"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: [string, "null"]
          format: date-time
        deletedAt:
          type: [string, "null"]
          format: date-time
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/DocumentSummary"
        - type: object
          required:
            [
              versions,
              storageObjects,
              relations,
              processingJobs,
              availableActions,
            ]
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/DocumentVersionSummary"
            storageObjects:
              type: array
              items:
                $ref: "#/components/schemas/DocumentStorageObject"
            relations:
              type: array
              items:
                $ref: "#/components/schemas/DocumentRelation"
            processingJobs:
              type: array
              items:
                $ref: "#/components/schemas/DocumentProcessingJob"
            availableActions:
              type: object
              required:
                [
                  canUploadVersion,
                  canDelete,
                  canRestore,
                  canManageTemplate,
                  canRequestSignedUrl,
                ]
              properties:
                canUploadVersion:
                  type: boolean
                canDelete:
                  type: boolean
                canRestore:
                  type: boolean
                canManageTemplate:
                  type: boolean
                canRequestSignedUrl:
                  type: boolean
    DocumentListResponse:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSummary"
        nextCursor:
          type: [string, "null"]
    DocumentUploadIntentRequest:
      type: object
      required:
        [title, kind, classification, originalFilename, mimeType, sizeBytes]
      properties:
        title:
          type: string
        description:
          type: string
        kind:
          $ref: "#/components/schemas/DocumentKind"
        classification:
          type: string
          enum:
            [
              public,
              internal,
              confidential,
              legal_secret,
              personal_data,
              client_material,
            ]
        originalFilename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        tags:
          type: array
          items:
            type: string
        relations:
          type: array
          items:
            type: object
            required: [relationType, targetEntityType, targetEntityId]
            properties:
              relationType:
                type: string
              targetEntityType:
                type: string
              targetEntityId:
                type: string
    CreateDocumentVersionUploadIntentRequest:
      type: object
      required: [originalFilename, mimeType, sizeBytes]
      properties:
        originalFilename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
    DocumentUploadIntentResponse:
      type: object
      required:
        [
          documentId,
          versionId,
          bucket,
          storagePath,
          uploadMethod,
          maxSizeBytes,
          allowedMimeTypes,
          expiresAt,
        ]
      properties:
        documentId:
          type: string
        versionId:
          type: string
        bucket:
          type: string
        storagePath:
          type: string
        uploadMethod:
          $ref: "#/components/schemas/UploadMethod"
        maxSizeBytes:
          type: integer
        allowedMimeTypes:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
    CompleteUploadRequest:
      type: object
      required: [clientReportedSize, clientReportedMimeType]
      properties:
        clientReportedSize:
          type: integer
        clientReportedMimeType:
          type: string
        sha256:
          type: [string, "null"]
    SignedUrlRequest:
      type: object
      required: [objectRole, purpose]
      properties:
        versionId:
          type: string
        objectRole:
          $ref: "#/components/schemas/DocumentObjectRole"
        purpose:
          type: string
          enum: [download, preview]
        expiresInSeconds:
          type: integer
    SignedUrlResponse:
      type: object
      required: [documentId, versionId, objectRole, signedUrl, expiresAt]
      properties:
        documentId:
          type: string
        versionId:
          type: string
        objectRole:
          $ref: "#/components/schemas/DocumentObjectRole"
        signedUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    DocumentMutationResult:
      type: object
      required: [status, documentId]
      properties:
        status:
          type: string
          enum: [archived, restored, deleted]
        documentId:
          type: string
...